requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
Django==4.2.9
djangorestframework==3.14.0
exceptiongroup==1.2.0
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
Markdown==3.5.2
packaging==23.2
pluggy==1.4.0
pytest==8.0.0
pytz==2023.4
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
zipp==3.17.0
